refactor(web): 抽离 SiteHealthBoard 组件供站点/监控总览共用 (#76)
