+++ frontend/package.json
@@ -21,6 +21,7 @@
   "devDependencies": {
     "@playwright/test": "^1.59.1",
     "@vitejs/plugin-vue": "^6.0.5",
+    "@vue/test-utils": "^2.4.11",
     "jsdom": "^29.0.2",
     "vite": "^8.0.4",
     "vitest": "^4.1.4"
