+++ frontend/vitest.config.js
@@ -1,7 +1,10 @@
 import { defineConfig } from 'vitest/config'
+import vue from '@vitejs/plugin-vue'
 
 export default defineConfig({
+  plugins: [vue()],
   test: {
+    environment: 'jsdom',
     exclude: ['e2e/**', 'node_modules/**'],
   },
 })
